pyproject.toml
[project]
name = "audition-client"
version = "0.1.0"
description = "The audition client application for Nanoko"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "keyring>=25.6.0",
    "nanoko-python>=0.1.1",
    "pyqt6>=6.9.0",
    "pyqt6-fluent-widgets>=1.8.1",
]

[dependency-groups]
dev = [
    "pydantic[email]>=2.11.4",
]
